Refactor DB connect & queries to use pooling, set up example route
