+++ server/config.js
@@ -1,5 +1,3 @@
-require('dotenv').config();
-
 module.exports = {
   SESSION_SECRET: process.env.SESSION_SECRET,
   PORT: process.env.PORT,
