+++ server/routes/example.js
@@ -1,7 +1,11 @@
 const router = require('express').Router();
+const db = require('../../db/queries.js');
 
 router.get('/', (req, res) => {
-  res.json({ data: ['dolphins', 'manatees', 'sea turles'] });
+  db.getUserData('Macey30@gmail.com').then(user => {
+    console.log('User is: ', user);
+    res.send(user);
+  });
 });
 
 module.exports = router;
